+++ app/controllers/RestApi.scala
@@ -25,7 +25,7 @@ import com.mohiva.play.silhouette.api.util.{PasswordHasher, PasswordInfo}
 import com.mohiva.play.silhouette.impl.providers.CredentialsProvider
 import com.mohiva.play.silhouette.impl.util.BCryptPasswordHasher
 import daos._
-import models.database.AccessRight
+import models.AccessRight
 import services.{TaskService, UserService, UserTokenService}
 import utils.Mailer
 import utils.Query.{QueryAST, QueryLexer, QueryParser}
