+++ app/daos/AccessRightDao.scala
@@ -4,7 +4,8 @@ import java.util.UUID
 import javax.inject.Inject
 
 import daos.schema._
-import models.database.AccessRight
+import models.AccessRight
+import models.database.AccessRightDB
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -38,19 +39,19 @@ class MariadbAccessRightDao @Inject()(dbConfigProvider: DatabaseConfigProvider) 
   val users = TableQuery[UserTableDef]
 
 
-  def all(): Future[Seq[AccessRight]] = dbConfig.db.run(accessRights.result)
+  def all(): Future[Seq[AccessRight]] = dbConfig.db.run(accessRights.result).map(aR => aR.map(accessRight => accessRight.toAccessRight))
 
   def forTask(taskId: Long) : Future[Seq[AccessRight]] = {
     val action = for {
       (ar, _) <- (accessRights join taskAccessRights.filter(tar => tar.taskId === taskId) on (_.id === _.accessRightId))
     } yield (ar)
-    dbConfig.db.run(action.result)
+    dbConfig.db.run(action.result).map(aR => aR.map(accessRight => accessRight.toAccessRight))
   }
 
-  def find(id: Long): Future[Option[AccessRight]] = dbConfig.db.run(accessRights.filter(ar => ar.id === id).result).map(_.headOption)
+  def find(id: Long): Future[Option[AccessRight]] = dbConfig.db.run(accessRights.filter(ar => ar.id === id).result).map(_.headOption.map(_.toAccessRight))
 
   def create(accessRight: AccessRight): Future[Option[AccessRight]] = {
-    dbConfig.db.run((accessRights returning accessRights.map(_.id)) += accessRight).flatMap((ac)=> find(ac))
+    dbConfig.db.run((accessRights returning accessRights.map(_.id)) += AccessRightDB(accessRight)).flatMap((ac)=> find(ac))
   }
 
   def delete(id: Long): Future[Int] = dbConfig.db.run(accessRights.filter(ar => ar.id === id).delete)
@@ -59,14 +60,14 @@ class MariadbAccessRightDao @Inject()(dbConfigProvider: DatabaseConfigProvider) 
     val action = for {
       (ar, _) <- (accessRights join taskAccessRights.filter(tar => tar.taskId.inSet(taskIdList)) on (_.id === _.accessRightId))
     } yield (ar)
-    dbConfig.db.run(action.result)
+    dbConfig.db.run(action.result).map(aR => aR.map(accessRight => accessRight.toAccessRight))
   }
 
   def forTaskListAndService(taskIdList : Seq[Long], service :String): Future[Seq[AccessRight]] = {
     val action = for {
       (ar, _) <- (accessRights.filter(aR => aR.service === service) join taskAccessRights.filter(tar => tar.taskId.inSet(taskIdList)) on (_.id === _.accessRightId))
     } yield (ar)
-    dbConfig.db.run(action.result)
+    dbConfig.db.run(action.result).map(aR => aR.map(accessRight => accessRight.toAccessRight))
   }
 
   def forUserAndService(userId: UUID, service: String): Future[Seq[AccessRight]] = {
@@ -80,7 +81,7 @@ class MariadbAccessRightDao @Inject()(dbConfigProvider: DatabaseConfigProvider) 
 
       } yield (ar)
 
-      dbConfig.db.run(action.result)
+      dbConfig.db.run(action.result).map(aR => aR.map(accessRight => accessRight.toAccessRight))
     })
   }
 }
